+++ CHANGELOG.markdown
@@ -8,6 +8,7 @@
 - Add Vibe/MultilineHashArgumentStyle cop.
 - Add Vibe/NoCompoundConditions cop.
 - Add Vibe/RspecBeforeBlockStyle cop.
+- Fix Vibe/BlankLineAfterAssignment to handle statement modifiers.
 
 ## 0.2.0 — January 8th, 2026